Move Vite-related functions to a single assets generator
